# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Urlytix API
  version: 0.1.0
  description: |
    Contract for implemented and planned Urlytix endpoints. Operations marked
    `x-urlytix-status: planned` are design targets and are not available yet.
    Billing is retained for a later phase and its implemented routes return 404
    unless the server is explicitly started with `BILLING_ENABLED=true`.
    Registering a user atomically provisions an OWNER default workspace.
    Browser sessions use API-managed httpOnly `access_token` (15 minutes) and
    rotating `refresh_token` (30 days) cookies; bearer JWTs remain supported.
    All paths are relative to the NestJS `/api` global prefix.
servers:
  - url: https://api.example.invalid/api
    description: Replace with the environment API origin
...
paths:
...
  /links:
    get:
      tags: [Links]
      operationId: listLinks
      summary: List links in a workspace
      x-urlytix-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkspaceIdQuery"
      responses:
        "200":
          description: Links
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Link"
                  meta:
                    type: object
                    required: [page, pageSize, total, totalPages]
                    properties:
                      page:
                        type: integer
                      pageSize:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Links]
      operationId: createLink
      summary: Create a short link
      x-urlytix-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLinkRequest"
      responses:
        "201":
          description: Link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /links/{linkId}:
    parameters:
      - $ref: "#/components/parameters/LinkId"
    get:
      tags: [Links]
      operationId: getLink
      summary: Get an owned link
      x-urlytix-status: implemented
      responses:
        "200":
          description: Link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Links]
      operationId: updateLink
      summary: Update an owned link
      x-urlytix-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLinkRequest"
      responses:
        "200":
          description: Updated link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Links]
      operationId: deleteLink
      summary: Delete an owned link
      x-urlytix-status: implemented
      responses:
        "200":
          description: Deletion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /links/{linkId}/status:
    parameters:
      - $ref: "#/components/parameters/LinkId"
    patch:
      tags: [Links]
      operationId: updateLinkStatus
      summary: Change link status
      x-urlytix-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLinkStatusRequest"
      responses:
        "200":
          description: Updated link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /r/{shortCode}:
    get:
      tags: [Links]
      operationId: redirectShortLink
      summary: Resolve and track a short link
      description: |
        Public short links are advertised as `/{shortCode}` on the API host
        (for example `https://go.example.com/abc1234`). That path is rewritten
        to this handler. The `/api/r/{shortCode}` form remains supported.
      x-urlytix-status: implemented
      security: []
      parameters:
        - name: shortCode
          in: path
          required: true
          schema:
            type: string
            minLength: 3
            maxLength: 32
      responses:
        "302":
          description: Redirect to destination
          headers:
            Location:
              schema:
                type: string
                format: uri
        "401":
          description: Password-protected link requires unlocking
          content:
            text/html:
              schema:
                type: string
        "403":
          description: Link is disabled
          content:
            text/html:
              schema:
                type: string
        "404":
          description: Link not found
          content:
            text/html:
              schema:
                type: string
        "410":
          description: Link expired
          content:
            text/html:
              schema:
                type: string
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      tags: [Links]
      operationId: unlockPasswordProtectedLink
      summary: Verify a protected-link password and redirect
      description: Limited to 5 attempts per minute per throttling identity.
      x-urlytix-status: implemented
      security: []
      parameters:
        - name: shortCode
          in: path
          required: true
          schema:
            type: string
            minLength: 3
            maxLength: 32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnlockLinkRequest"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/UnlockLinkRequest"
      responses:
        "302":
          description: Password accepted; redirect to destination
          headers:
            Location:
              schema:
                type: string
                format: uri
            Set-Cookie:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Password is required or invalid
          content:
            text/html:
              schema:
                type: string
        "404":
          description: Link not found
          content:
            text/html:
              schema:
                type: string
        "410":
          description: Link expired
          content:
            text/html:
              schema:
                type: string
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    accessCookie:
      type: apiKey
      in: cookie
      name: access_token
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
  parameters:
    LinkId:
      name: linkId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    WorkspaceIdQuery:
      name: workspaceId
      in: query
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Resource state conflicts with the request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [statusCode, message]
      properties:
        statusCode:
          type: integer
          minimum: 400
          maximum: 599
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        error:
          type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
    Message:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    LinkStatus:
      type: string
      enum: [ACTIVE, DISABLED, EXPIRED]
    CreateLinkRequest:
      type: object
      additionalProperties: false
      required: [workspaceId, originalUrl]
      properties:
        workspaceId:
          type: string
          format: uuid
        originalUrl:
          type: string
          format: uri
        title:
          type: string
          maxLength: 80
        customAlias:
          type: string
          minLength: 3
          maxLength: 32
          pattern: "^[a-zA-Z0-9_-]+$"
        expiresAt:
          type: string
          format: date-time
        password:
          type: string
          minLength: 4
          maxLength: 72
    UpdateLinkRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        originalUrl:
          type: string
          format: uri
        title:
          type: string
          maxLength: 80
        customAlias:
          type: string
          minLength: 3
          maxLength: 32
          pattern: '^[a-zA-Z0-9_-]+$'
          description: Changes the public short path. The previous short URL stops working.
        expiresAt:
          type: [string, "null"]
          format: date-time
        password:
          type: [string, "null"]
          minLength: 4
          maxLength: 72
    UnlockLinkRequest:
      type: object
      additionalProperties: false
      required: [password]
      properties:
        password:
          type: string
          minLength: 1
          maxLength: 72
    UpdateLinkStatusRequest:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/LinkStatus"
    Link:
      type: object
      required: [id, originalUrl, shortCode, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        originalUrl:
          type: string
          format: uri
        shortCode:
          type: string
        shortUrl:
          type: string
          format: uri
        title:
          type: [string, "null"]
        status:
          $ref: "#/components/schemas/LinkStatus"
        expiresAt:
          type: [string, "null"]
          format: date-time
        hasPassword:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
  - accessCookie: []